fix: separate agenda into own service
