--- src/background.html
@@ -1,1 +1,0 @@
-<script type="module" src="js/background.js"></script>
